Folders and files in new convention. Need to destruct into components.
