--- components/selectors/Selectors.js
@@ -1,24 +1,0 @@
-import React, { Component } from 'react'
-
-class Selectors extends Component {
-
-  constructor(props, context) {
-    super(props, context)
-    this.state = {
-      id: 0,
-      text: 'Stuff',
-      done: false
-    }
-  }
-
-  render() {
-    return (
-      <div>
-
-      </div>
-    )
-  }
-
-}
-
-export default Selectors
