+++ package.json
@@ -25,7 +25,7 @@
     "webpack-hot-middleware": "^2.6.4"
   },
   "scripts": {
-    "start": "nodemon server/Server.js --ignore components",
+    "start": "nodemon server/server.js --ignore components",
     "build": "react-scripts build",
     "test": "react-scripts test --env=jsdom"
   },
