+++ redux/ConfigureStore.js
@@ -1,13 +1,13 @@
 import { applyMiddleware, compose, createStore } from 'redux'
 import logger from 'redux-logger'
 import thunk from 'redux-thunk'
-import rootReducer from './RootReducer'
+import rootReducer from './reducers/rootReducer'
 
 let finalCreateStore = compose(
   applyMiddleware(thunk, logger)
 )(createStore)
 
 
-export default function configureStore(initialState = {workOrder: {}, itemTest: []}) {
+export default function configureStore(initialState = {workOrder: {}, itemTest: [], history: [], instructions: []}) {
   return finalCreateStore(rootReducer, initialState)
 }
