+++ redux/reducers/rootReducer.js
@@ -1,6 +1,6 @@
 import { combineReducers } from 'redux'
-import SelectorReducer from './SelectorReducer'
-import ModulesReducer from './ModulesReducer'
+import SelectorReducer from './selectorReducer'
+import ModulesReducer from './modulesReducer'
 
 export default combineReducers({
   workOrder: ModulesReducer,
