+++ server/Server.js
@@ -15,7 +15,7 @@ app.use(webpackHotMiddleware(compiler));
 app.use(express.static('./dist'));
 
 app.use('/', function (req, res) {
-  res.sendFile(path.resolve('client/index.html'));
+  res.sendFile(path.resolve('public/index.html'));
 });
 
 var port = 3000;
