--- styles/modules.css
@@ -1,5 +1,0 @@
-.modules {
-  padding: 0;
-  background-color: #E5FFCC;
-  flex: 1;
-}
