+++ styles/sideBar.css
@@ -1,4 +1,4 @@
-.sideMenu {
+.sideBar {
   margin: 0;
   padding: 0;
   width: 200px;
@@ -8,19 +8,19 @@
   border-right: 1px solid #e4e4e4;
 }
 
-.sideMenu hr {
+.sideBar hr {
   height: 0;
   margin: 0;
   border: 0;
   border-bottom: 1px solid rgba(0, 0, 0, 0.12);
 }
 
-.sideMenu section {
+.sideBar section {
   display: flex;
   flex-flow: column nowrap;
 }
 
-.sideMenu section a {
+.sideBar section a {
   padding: 12px 16px;
   text-decoration: none;
   font-weight: 500;
@@ -30,11 +30,11 @@
   flex-flow: row nowrap;
 }
 
-.sideMenu section a span {
+.sideBar section a span {
   align-self: center;
 }
 
-.sideMenu section a img {
+.sideBar section a img {
   width: 24px;
   height: 24px;
   margin-right: 32px;
